Address review comments, remove iter file
